# Auto-flush for all channels.