--- a/common/src/main/java/io/netty/util/LongBiConsumer.java
+++ b/common/src/main/java/io/netty/util/LongBiConsumer.java
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2016 The Netty Project
+ *
+ * The Netty Project licenses this file to you under the Apache License,
+ * version 2.0 (the "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at:
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ */
+package io.netty.util;
+
+/**
+ * Consumer of two {@code long} values.
+ */
+public interface LongBiConsumer {
+
+    /**
+     * Accepts the passed {@code long} values.
+     *
+     * @param aLong1 to accept.
+     * @param aLong2 to accept.
+     */
+    void accept(long aLong1, long aLong2);
+}