--- a/transport/src/main/java/io/netty/channel/ChannelOption.java
+++ b/transport/src/main/java/io/netty/channel/ChannelOption.java
@@ -18,6 +18,7 @@ package io.netty.channel;
 import io.netty.buffer.ByteBufAllocator;
 import io.netty.util.AbstractConstant;
 import io.netty.util.ConstantPool;
+import io.netty.util.internal.UnstableApi;
 
 import java.net.InetAddress;
 import java.net.NetworkInterface;
@@ -97,6 +98,15 @@ public class ChannelOption<T> extends AbstractConstant<ChannelOption<T>> {
 
     public static final ChannelOption<Boolean> ALLOW_HALF_CLOSURE = valueOf("ALLOW_HALF_CLOSURE");
     public static final ChannelOption<Boolean> AUTO_READ = valueOf("AUTO_READ");
+
+    /**
+     * Provides a way to eliminate explicit flushing of writes on this channel. If turned-on, any pending writes on the
+     * channel will be flushed on the next iteration of the eventloop if possible or directly (depending on the
+     * implementation).
+     * Explicit flushes will still be honored.
+     */
+    @UnstableApi
+    public static final ChannelOption<Boolean> AUTO_FLUSH = valueOf("AUTO_FLUSH");
 
     /**
      * @deprecated From version 5.0, {@link Channel} will not be closed on write failure.