###   Motivation

  Netty is extremely flexible in providing a way to batch writes on the physical socket by providing a means to `write` and `flush` writes separately. This works for the usecases when the application pattern is one that provides predictable finite streams of data. In cases, when an infinite stream of data is to be written on a channel which does not have any predictable length and duration of writes, a user has the following choices:

 - Flush every write. 
 - Flush batches based on time and count (eg: Flush every 10 message or 1 second)   

Neither of the above is ideal as the nature of stream is completely unpredictable. Flushing every write is costly as `flush` is a system call. Flushing batches is arbitrary and has to be deviced for each stream in an arbitrary way.  

### Modifications  

This change is a Proof Of Concept, is __not__ complete and has the following missing pieces:

  1. _UnitTests_: I haven't yet added any tests.
2. `DefaultChannelPipeline` caches the value of `ChannelConfig.isAutoRead()` at channel activation and does not update it later.
3. _Documentation_: I haven’t yet added any javadocs or comments to explain the changes.   

__I would like to invest more time on this change, if there is some consensus on whether this makes sense.  __

 Here are the details of the change:   

##### `ChannelOption.AUTO_FLUSH`   

 A channel option for enable/disable auto flush (this feature) is added.  When, this option is enabled, explicit flushes of writes on that channel are not required. Users can just call `write` and these writes will be flushed on every iteration of the associated `eventLoop` for that channel.
Any explicit flushes on the channel work as before i.e. it flushes all writes that happened before the flush and were not flushed. 

##### Changes to `EventLoop`

Two methods:

```java
void onEventLoopIteration(Runnable task);  
boolean removeOnEventLoopIterationTask(Runnable task);
```

are added to `EventLoop` interface for adding/removing a task to be executed at the end of each iteration of this `eventloop`.

##### Changes to `SingleThreadEventExecutor`

Added a method to `SingleThreadEventExecutor`

```java
protected void afterRunningAllTasks(boolean timedOut) { }

```

This is invoked after all tasks are run for this executor OR if the passed timeout value for `runAllTasks(long timeoutNanos)` is expired. In case of `timeout` the argument to this method is `true`.

##### Changes to `SingleThreadEventLoop`

Added a queue of `tailTasks` to this class to hold all tasks to be executed at the end of every iteration.
In the case if `afterRunningAllTasks` is invoked as a result of a timeout, no tail tasks are executed. Instead, the tail tasks are executed at the start of the next iteration.

##### Changes to `DefaultChannelPipeline`

Additional calls to `wakeup` the selector for the first `unflushed` write. This is done so that there isn’t a latency between the first `unflushed` write and invocation of the tails tasks.

##### Changes to `AbstractChannel`

Manages the lifecycle of `AutoFlushTask` per channel as the task to be invoked per iteration of the eventloop.
 
### Result

  This results in a much better model for infinite streams written to a channel.

I wrote a micro-benchmark(`io.netty.microbench.channel.AutoFlushBenchmark`) to demonstrate the throughput variations between all possible current alternatives.
Here are the results for the same:

```
Benchmark                                             (flush)  (writeCount)   Mode  Cnt      Score       Error  Units
AutoFlushBenchmark.measureWritesWithFlushAtEnd           true             1  thrpt   10  50679.715 ±  1557.854  ops/s
AutoFlushBenchmark.measureWritesWithFlushAtEnd           true            10  thrpt   10  38835.955 ±   887.623  ops/s
AutoFlushBenchmark.measureWritesWithFlushAtEnd           true           100  thrpt   10  11916.433 ±    96.373  ops/s
AutoFlushBenchmark.measureWritesWithFlushAtEnd          false             1  thrpt   10  50452.022 ±  1366.724  ops/s
AutoFlushBenchmark.measureWritesWithFlushAtEnd          false            10  thrpt   10  37895.350 ±   492.081  ops/s
AutoFlushBenchmark.measureWritesWithFlushAtEnd          false           100  thrpt   10  11586.364 ±   132.067  ops/s
AutoFlushBenchmark.measureWritesWithFlushEvery5          true             1  thrpt   10  50592.297 ±  1102.360  ops/s
AutoFlushBenchmark.measureWritesWithFlushEvery5          true            10  thrpt   10  27176.931 ±   335.812  ops/s
AutoFlushBenchmark.measureWritesWithFlushEvery5          true           100  thrpt   10   5447.520 ±    82.222  ops/s
AutoFlushBenchmark.measureWritesWithFlushEvery5         false             1  thrpt   10  49741.284 ±   752.898  ops/s
AutoFlushBenchmark.measureWritesWithFlushEvery5         false            10  thrpt   10  37140.030 ±  1556.103  ops/s
AutoFlushBenchmark.measureWritesWithFlushEvery5         false           100  thrpt   10  11503.099 ±    94.868  ops/s
AutoFlushBenchmark.measureWritesWithFlushEverySecond     true             1  thrpt   10  36295.567 ± 13013.318  ops/s
AutoFlushBenchmark.measureWritesWithFlushEverySecond     true            10  thrpt   10  26575.064 ±  3091.255  ops/s
AutoFlushBenchmark.measureWritesWithFlushEverySecond     true           100  thrpt   10   8088.351 ±  1390.408  ops/s
AutoFlushBenchmark.measureWritesWithFlushEverySecond    false             1  thrpt   10  49721.904 ±   718.826  ops/s
AutoFlushBenchmark.measureWritesWithFlushEverySecond    false            10  thrpt   10  38451.303 ±   712.448  ops/s
AutoFlushBenchmark.measureWritesWithFlushEverySecond    false           100  thrpt   10  11387.047 ±   139.104  ops/s
AutoFlushBenchmark.measureWritesWithFlushOnEach          true             1  thrpt   10  51501.154 ±  3866.709  ops/s
AutoFlushBenchmark.measureWritesWithFlushOnEach          true            10  thrpt   10  15308.538 ±   351.524  ops/s
AutoFlushBenchmark.measureWritesWithFlushOnEach          true           100  thrpt   10   1746.113 ±    31.679  ops/s
AutoFlushBenchmark.measureWritesWithFlushOnEach         false             1  thrpt   10  49927.340 ±  1482.713  ops/s
AutoFlushBenchmark.measureWritesWithFlushOnEach         false            10  thrpt   10  37389.582 ±  1859.288  ops/s
AutoFlushBenchmark.measureWritesWithFlushOnEach         false           100  thrpt   10  11604.148 ±   342.597  ops/s
```

This was done on MBP with 16GB of memory and 8 processors.