@normanmaurer 2016-06-15
@NiteshKant did not have a look at the code yet but one "limitation" that comes into my mind is that it will only work if you ever only will write from within the EventLoop. Otherwise not.