@normanmaurer 2016-06-15
@NiteshKant maybe in this case a user calls write from outside the EventLoop it would be good enough to always just flush on write.