@johnou 2016-06-15
@normanmaurer can you elaborate on outside of the EventLoop? My client session looks something like this.. 

```java
    public void sendToClient(OutgoingClientMessage message, boolean close) {
        safeSendToClient(message.getByteBuf(), close);
    }

    private void safeSendToClient(ByteBuf buffer, boolean close) {
        if (channel.eventLoop().inEventLoop()) {
            sendToClient(buffer, close);
        } else {
            try {
                channel.eventLoop().execute(() -> sendToClient(buffer, close));
            } catch (RuntimeException e) {
                logger.error("Failed sending payload to client", e);
                ReferenceCountUtil.safeRelease(buffer);
            }
        }
    }

    private void sendToClient(ByteBuf buffer, boolean close) {
        if (close) {
            channel.config().setAutoRead(false);
            ChannelFuture future = channel.writeAndFlush(buffer);
            future.addListener(ChannelFutureListener.CLOSE);
        } else {
            channel.writeAndFlush(buffer, channel.voidPromise());
        }
    }
```

but on looking through the Netty writeAndFlush code it seems to be that the task would always be written inside the eventloop?

```java

```