@NiteshKant 2016-06-15
@normanmaurer If I understand you correctly, what you are saying is that it is "effective" only if write happens from within eventloop i.e. it will be able to effectively batch more writes that way. Is that what you meant?

For writes from within eventloop, for most cases (wherein write happens after read from the channel) having `flush` inside `onReadComplete()` is much more effective as it is guaranteed to batch all writes from within reads. Having said that, one can remove that pattern (`flush()` in `onReadComplete()`) with this option and have very similar behavior.

I believe this option is much more useful when writes happen from outside the eventloop and for multiplexed connections (eg: HTTP/2). Let us take an example of HTTP/2 which has 5 concurrent requests that are to be written, each of which consists of 1-3 frames. Typically, one will use `writeAndFlush()` or `write()`*n + `flush()`(for each disparate request), in which case eventloop task queue will look something like this:

```
W W F W F W W W F W F W F
```

Assuming that these writes are really concurrent, they will most likely happen before the eventloop runs through all existing tasks. If that is the case, then by using this option, the task queue will look like:

```
W W W W W W W W F
```

Which in effect de-dups the multiple flushes and replace them with a single flush. The first write here will `wakeup` the eventloop and by the time it runs through its task list, it would have seen more writes. OTOH, if the eventloop is already running then it depends on the timing as to how many flushes will be invoked. The worst case would be equivalent to `writeAndFlush()` and the best case is what I described above. The benchmark I have added in this PR (results in the main description) gives some hint of this behavior although they are not very clear as to how many writes were batched.