@normanmaurer 2016-06-15
@NiteshKant the problem is that a write from outside the EventLoop will NOT wakeup the Selector and so not cause a new EventLoop run. That will only happen on a flush() for performance reasons.