@normanmaurer 2016-06-15
@NiteshKant had a first look at the code. Look promising!