@NiteshKant 2016-06-16
@normanmaurer thanks for the quick review! 
So, looks like the idea is acceptable. I will address the review comments and the other open items I mentioned before.

(I think you noticed this in the code, but for the benefit of others following this PR)

>> the problem is that a write from outside the EventLoop will NOT wakeup the Selector and so not cause a new EventLoop run.

I `wakeup` the selector for the first write on a channel between two runs of the auto-flush task.