@NiteshKant 2016-06-19
@normanmaurer This is ready for another round of reviews. I have addressed the TODOs from my end and all review comments. The only pending item is your suggestion about moving the wakeup code for selector to the `AbstractChannelHandlerContext`; I will wait for your reply to my reasons above.