@normanmaurer 2016-06-20
@NiteshKant id another round of review...

@trustin PTAL as well.