@NiteshKant 2016-06-21
@normanmaurer I added a benchmark to measure writes. It measures write throughput with 1, 5 and 10 handlers in the pipeline and writing 1 and 10 messages. One test flushes on every write (this doesn't touch the volatile flag for autoFlush) and other flushes at the end of all writes. Below are the numbers:

#### With these changes

```
Benchmark                                          (handlerCount)  (writeCount)   Mode  Cnt      Score      Error  Units
ChannelWriteBenchmark.measureWriteWithFlushAtEnd                1             1  thrpt   10  42599.250 ± 3270.590  ops/s
ChannelWriteBenchmark.measureWriteWithFlushAtEnd                1            10  thrpt   10  35981.618 ± 3397.484  ops/s
ChannelWriteBenchmark.measureWriteWithFlushAtEnd                5             1  thrpt   10  41773.690 ± 2407.105  ops/s
ChannelWriteBenchmark.measureWriteWithFlushAtEnd                5            10  thrpt   10  35578.091 ± 5378.863  ops/s
ChannelWriteBenchmark.measureWriteWithFlushAtEnd               10             1  thrpt   10  44164.664 ± 4700.700  ops/s
ChannelWriteBenchmark.measureWriteWithFlushAtEnd               10            10  thrpt   10  36966.154 ±  552.128  ops/s
ChannelWriteBenchmark.measureWriteWithFlushOnEach               1             1  thrpt   10  46531.750 ± 3655.686  ops/s
ChannelWriteBenchmark.measureWriteWithFlushOnEach               1            10  thrpt   10  12010.405 ±  981.604  ops/s
ChannelWriteBenchmark.measureWriteWithFlushOnEach               5             1  thrpt   10  47078.183 ± 1017.086  ops/s
ChannelWriteBenchmark.measureWriteWithFlushOnEach               5            10  thrpt   10  12336.481 ±  518.871  ops/s
ChannelWriteBenchmark.measureWriteWithFlushOnEach              10             1  thrpt   10  46628.070 ± 1242.385  ops/s
ChannelWriteBenchmark.measureWriteWithFlushOnEach              10            10  thrpt   10  12078.870 ±  181.724  ops/s
```

#### Without these changes

```
Benchmark                                          (handlerCount)  (writeCount)   Mode  Cnt      Score      Error  Units
ChannelWriteBenchmark.measureWriteWithFlushAtEnd                1             1  thrpt   10  47062.062 ±  552.498  ops/s
ChannelWriteBenchmark.measureWriteWithFlushAtEnd                1            10  thrpt   10  37382.617 ±  581.512  ops/s
ChannelWriteBenchmark.measureWriteWithFlushAtEnd                5             1  thrpt   10  46696.305 ±  667.436  ops/s
ChannelWriteBenchmark.measureWriteWithFlushAtEnd                5            10  thrpt   10  37506.526 ±  376.604  ops/s
ChannelWriteBenchmark.measureWriteWithFlushAtEnd               10             1  thrpt   10  46786.772 ±  835.597  ops/s
ChannelWriteBenchmark.measureWriteWithFlushAtEnd               10            10  thrpt   10  35733.629 ± 1497.456  ops/s
ChannelWriteBenchmark.measureWriteWithFlushOnEach               1             1  thrpt   10  44659.447 ± 4275.720  ops/s
ChannelWriteBenchmark.measureWriteWithFlushOnEach               1            10  thrpt   10  12421.550 ±  809.539  ops/s
ChannelWriteBenchmark.measureWriteWithFlushOnEach               5             1  thrpt   10  44118.946 ± 1844.873  ops/s
ChannelWriteBenchmark.measureWriteWithFlushOnEach               5            10  thrpt   10  12449.186 ±  776.835  ops/s
ChannelWriteBenchmark.measureWriteWithFlushOnEach              10             1  thrpt   10  45209.271 ± 2730.433  ops/s
ChannelWriteBenchmark.measureWriteWithFlushOnEach              10            10  thrpt   10  11972.691 ±  370.274  ops/s
```

WDYT?